requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==2.0.1
openpyxl==3.1.5
XlsxWriter==3.2.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
